Eliminate repetative setup from tests
Instead of setting up http mock on each test, we can do so in TestMain
This will be the first of a series of cleanups to make the tests
easier to understand and fuller
