+++ decks_test.go
@@ -5,20 +5,12 @@ import (
 	"testing"
 
 	"github.com/jarcoal/httpmock"
-	"github.com/privatesquare/bkst-go-utils/utils/fileutils"
 	"github.com/stretchr/testify/assert"
-)
-
-const (
-	testDataPath          = "data/test/"
-	errorTestDataFileName = "error.json"
-	jsonExt               = ".json"
 )
 
 func TestDecksManager_GetAll(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		result := new(Result[[]string])
 		loadTestData(t, testDataPath+ActionDeckNames+jsonExt, result)
@@ -34,8 +26,7 @@ func TestDecksManager_GetAll(t *testing.T) {
 	})
 
 	t.Run("error", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		result := new(Result[[]string])
 		loadTestData(t, testDataPath+errorTestDataFileName, result)
@@ -52,8 +43,7 @@ func TestDecksManager_GetAll(t *testing.T) {
 	})
 
 	t.Run("http request error", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		decks, restErr := client.Decks.GetAll()
 		assert.Nil(t, decks)
@@ -65,8 +55,7 @@ func TestDecksManager_GetAll(t *testing.T) {
 
 func TestDecksManager_Create(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		result := new(Result[int64])
 		loadTestData(t, testDataPath+ActionCreateDeck+jsonExt, result)
@@ -80,8 +69,7 @@ func TestDecksManager_Create(t *testing.T) {
 	})
 
 	t.Run("error", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		result := new(Result[string])
 		loadTestData(t, testDataPath+errorTestDataFileName, result)
@@ -99,8 +87,7 @@ func TestDecksManager_Create(t *testing.T) {
 
 func TestDecksManagerDelete(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		result := new(Result[string])
 		loadTestData(t, testDataPath+ActionDeleteDecks+jsonExt, result)
@@ -114,8 +101,7 @@ func TestDecksManagerDelete(t *testing.T) {
 	})
 
 	t.Run("error", func(t *testing.T) {
-		httpmock.ActivateNonDefault(client.httpClient.GetClient())
-		defer httpmock.DeactivateAndReset()
+		defer httpmock.Reset()
 
 		result := new(Result[string])
 		loadTestData(t, testDataPath+errorTestDataFileName, result)
@@ -131,7 +117,3 @@ func TestDecksManagerDelete(t *testing.T) {
 	})
 }
 
-func loadTestData(t *testing.T, path string, out interface{}) {
-	err := fileutils.ReadJsonFile(path, &out)
-	assert.NoError(t, err)
-}
